feat: implement PostModel and add post feature with like and comment
This commit introduces the `PostModel` to manage post data efficiently. It implements the functionality to add new posts to the Firestore database, including title, caption, imageUrls, author username, group, and creation timestamp. Additionally, it allows users to like posts and add replies, which are stored in a sub-collection within each post document.
 - Create new post model file
 - Add `title`, `caption`, `imageUrls`, `authorUsername`, `group`, `likes`, `createdAt`, `likedBy` to `PostModel`
 - Add `addPost` method to `FirestoreDatabase`
 - Add create post dialog to `home_page.dart`
 - Add `toggleLike` method to `FirestoreDatabase` to update likes and likedBy field
 - Add `addReply` method to `FirestoreDatabase` to add comments
 - Add `post_detail_page.dart` to show post detail and comment
 - Show posts in `home_page.dart` with like and reply button